Moved MapzenVectorTileJson to its own
and related changes
