+++ core/src/data/dataSource.h
@@ -69,27 +69,11 @@ public:
 
 };
 
-/* Extends NetworkDataSource class to read Mapzen's GeoJSON vector tiles */
-class MapzenVectorTileJson: public NetworkDataSource {
-
-protected:
-    
-    virtual std::shared_ptr<TileData> parse(const MapTile& _tile, std::stringstream& _in) override;
-    
-public:
-
-    MapzenVectorTileJson();
-
-};
-
 // TODO: Support TopoJSON tiles
 class TopoJsonNetSrc : public NetworkDataSource {
-};
-
-// TODO: Support Mapbox tiles
-class MapboxFormatNetSrc : public NetworkDataSource {
 };
 
 // TODO: Support local GeoJSON tiles
 class GeoJsonFileSrc : public DataSource {
 };
+
